tratando tópicos que não existem disparando status code 404

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -32,13 +33,13 @@ public class TopicosController {
 
 	@Autowired
 	private TopicoRepository topicoRepository;
-	
+
 	@Autowired
 	private CursoRepository cursoRepository;
-	
+
 	@GetMapping
 	public List<TopicoDto> lista(String nomeCurso) {
-		if (nomeCurso == null || nomeCurso == "") {			
+		if (nomeCurso == null || nomeCurso == "") {
 			List<Topico> topicos = topicoRepository.findAll();
 			return TopicoDto.converter(topicos);
 		}
@@ -46,33 +47,48 @@ public class TopicosController {
 		List<Topico> topicos = topicoRepository.findByCursoNome(nomeCurso);
 		return TopicoDto.converter(topicos);
 	}
-	
+
 	@PostMapping
 	@Transactional
 	public ResponseEntity<TopicoDto> cadastrar(@RequestBody @Valid TopicoForm form, UriComponentsBuilder uriBuilder) {
 		Topico topico = form.converter(cursoRepository);
 		topicoRepository.save(topico);
-		
+
 		URI uri = uriBuilder.path("/topicos/{id}").buildAndExpand(topico.getId()).toUri();
-		return ResponseEntity.created(uri).body(new TopicoDto(topico)); 
+		return ResponseEntity.created(uri).body(new TopicoDto(topico));
 	}
-	
+
 	@GetMapping("/{id}")
-	public DetalhesTopicoDto detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalhesTopicoDto(topico);
+	public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if (topico.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		}
+
+		return ResponseEntity.ok(new DetalhesTopicoDto(topico.get()));
 	}
-	
+
 	@PutMapping("/{id}")
 	@Transactional
 	public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if (optional.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		}
+		
 		Topico topico = form.atualizar(id, topicoRepository);
 		return ResponseEntity.ok(new TopicoDto(topico));
 	}
-	
+
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> remover(@PathVariable Long id) {
+		Optional<Topico> optional = topicoRepository.findById(id);
+
+		if (optional.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		}
+		
 		topicoRepository.deleteById(id);
 		return ResponseEntity.ok().build();
 	}
